Fix dependency download failure. Use latest version of Dispatch and specify the Scala version we want to use.
SBT failed to resolve a dependency (the dependency was related to scalaz) due to the resolver for scalaz not being present in this file, so it was added. In order to prevent future strangeness, I chose to use the latest integration of Dispatch and specify exactly the Scala version we're using.